+++ web/src/lib/features/ticket-detail/types.ts
@@ -1,4 +1,7 @@
 import type { ChatDiffPayload } from '$lib/api/chat'
+import type { TicketPickupDiagnosis } from './pickup-diagnosis'
+
+export type { TicketPickupDiagnosis } from './pickup-diagnosis'
 
 export type TicketStatusOption = {
   id: string
@@ -35,57 +38,6 @@ export type TicketExternalLinkDraft = {
   title: string
   status: string
   relation: string
-}
-
-export type TicketPickupDiagnosis = {
-  state: 'runnable' | 'waiting' | 'blocked' | 'running' | 'completed' | 'unavailable'
-  primaryReasonCode: string
-  primaryReasonMessage: string
-  nextActionHint?: string
-  reasons: Array<{
-    code: string
-    message: string
-    severity: 'info' | 'warning' | 'error'
-  }>
-  workflow?: {
-    id: string
-    name: string
-    isActive: boolean
-    pickupStatusMatch: boolean
-  }
-  agent?: {
-    id: string
-    name: string
-    runtimeControlState: string
-  }
-  provider?: {
-    id: string
-    name: string
-    machineId: string
-    machineName: string
-    machineStatus: string
-    availabilityState: string
-    availabilityReason?: string
-  }
-  retry: {
-    attemptCount: number
-    retryPaused: boolean
-    pauseReason?: string
-    nextRetryAt?: string
-  }
-  capacity: {
-    workflow: { limited: boolean; activeRuns: number; capacity: number }
-    project: { limited: boolean; activeRuns: number; capacity: number }
-    provider: { limited: boolean; activeRuns: number; capacity: number }
-    status: { limited: boolean; activeRuns: number; capacity?: number }
-  }
-  blockedBy: Array<{
-    id: string
-    identifier: string
-    title: string
-    statusId: string
-    statusName: string
-  }>
 }
 
 export type TicketDetail = {
